+++ cmd/infra_secrets.go
@@ -70,14 +70,12 @@ func getSecrets(kubectlPath string, verboseFlag bool) error {
 	}
 
 	openfaasPassword, _ := base64.StdEncoding.DecodeString(res.Stdout)
-	fmt.Println("\nOpenFaaS - gateway.ipaas.localdomain:8080")
+	fmt.Println("OpenFaaS - gateway.ipaas.localdomain:8080")
 	fmt.Println("    User: admin")
 	fmt.Println("Password: " + string(openfaasPassword))
-	fmt.Println()
 	fmt.Println("\nKonga - konga.localdomain:8080")
 	fmt.Println("    User: admin")
 	fmt.Println("Password: admin123")
-	fmt.Println()
 
 	return nil
 }